Move readFile and execCommand into their own platform files
